[GPU] Add SwapExpandShapeWithSlice pattern to loop fusion pass
