+++ compiler/src/iree/compiler/Codegen/Common/BUILD.bazel
@@ -155,6 +155,7 @@ iree_compiler_cc_library(
         "TileDispatchUsingInterface.cpp",
         "TileLargeTensors.cpp",
         "TileSizeSelection.cpp",
+        "Transforms.cpp",
         "TypePropagationPass.cpp",
         "UnrollAnnotatedLoops.cpp",
         "UserConfig.cpp",
